// types/editor-language.ts
export interface EditorLanguage {
    LanguageName: "Javascript" | "Java" | "Python" | "C" | "C++" | "C#" | "Golang" | "Rust" | "Kotlin" | "Typescript";
    EditorLanguage: "javascript" | "java" | "python" | "c_cpp" | "csharp" | "golang" | "kotlin" | "rust" | "typescript";
    APILanguage: "js" | "java" | "py" | "c" | "cpp" | "cs" | "go" | "rs" | "kt" | "ts";
}